refactor: okhttp 사용하는 데이터소스 이름 service로 변경
