+++ app/src/main/java/woowacourse/shopping/ui/shopping/ShoppingActivity.kt
@@ -14,8 +14,8 @@ import woowacourse.shopping.model.ProductUIModel
 import woowacourse.shopping.model.RecentProductUIModel
 import woowacourse.shopping.repositoryImpl.CartRepositoryImpl
 import woowacourse.shopping.repositoryImpl.ProductRepositoryImpl
-import woowacourse.shopping.repositoryImpl.RemoteCartDataSource
-import woowacourse.shopping.repositoryImpl.RemoteProductDataSource
+import woowacourse.shopping.service.RemoteCartService
+import woowacourse.shopping.service.RemoteProductService
 import woowacourse.shopping.ui.cart.CartActivity
 import woowacourse.shopping.ui.detailedProduct.DetailedProductActivity
 import woowacourse.shopping.ui.shopping.productAdapter.ProductsAdapter
@@ -75,10 +75,10 @@ class ShoppingActivity : AppCompatActivity(), ShoppingContract.View {
             this,
             ProductRepositoryImpl(
                 ProductDatabase(this),
-                RemoteProductDataSource(ServerURL.url)
+                RemoteProductService(ServerURL.url)
             ),
             RecentProductDatabase(this),
-            CartRepositoryImpl(RemoteCartDataSource(ServerURL.url))
+            CartRepositoryImpl(RemoteCartService(ServerURL.url))
         )
         presenter.setUpRecentProducts()
         presenter.setUpNextProducts()
